refactor: Move read file to handler layer
